+++ src/presenters/includes/team-profile.jsx
@@ -3,9 +3,26 @@ import PropTypes from 'prop-types';
 
 import Loader from '../includes/loader.jsx';
 import Thanks from './thanks.jsx';
-import TeamUsers from "../includes/team-users.jsx";
 import AddTeamUser from '../includes/add-team-user.jsx';
 import {EditableDescription, StaticDescription} from './description-field.jsx';
+import UserInfoPop from '../pop-overs/user-info-pop.jsx';
+import {UserPopoversList} from '../users-list.jsx';
+
+const TeamUsers = ({users, currentUserIsOnTeam, removeUserFromTeam}) => (
+  <UserPopoversList users={users}>
+    {user => <UserInfoPop user={user} currentUserIsOnTeam={currentUserIsOnTeam} removeUserFromTeam={() => removeUserFromTeam(user)} />}
+  </UserPopoversList>
+);
+
+TeamUsers.propTypes = {
+  users: PropTypes.array.isRequired,
+  currentUserIsOnTeam: PropTypes.bool.isRequired,
+  removeUserFromTeam: PropTypes.func.isRequired,
+};
+
+export default TeamUsers;
+
+//...
 
 const UserAvatarContainer = ({
   addUserToTeam,
